fix consistent prettier config resolution
