+++ .changeset/plenty-impalas-double.md
@@ -1,0 +1,5 @@
+---
+'@graphql-ez/plugin-codegen': patch
+---
+
+Fix consistent prettier config resolution
